Removed futil, put functionality inside builder
